Preload Rails’ route helpers into module
This means the Rails app needs to be declared before requiring this gem, so

gem `zooplankton`, require: false

And then `require “zooplankton”` in an initializer.
