+++ lib/zooplankton/resolver.rb
@@ -1,6 +1,22 @@
+require "rails"
 require 'action_controller/metal/strong_parameters'
 
 module Zooplankton
+  # we need to preload routes and helpers
+  # to avoid this threading issue https://github.com/puma/puma/issues/647
+  # routes.url_helpers creates a module every time, which is slow
+  # and should be done only once on app startup.
+  module Routes
+    module UrlHelpers
+      include Rails.application.routes.url_helpers
+    end
+    extend UrlHelpers
+
+    def self.default_url_options
+      Rails.application.routes.default_url_options
+    end
+  end
+
   class Resolver
     def self.instance
       new(Rails.application.routes)
@@ -16,7 +32,7 @@ module Zooplankton
 
     def generate(helper_method, helper_name, params = {})
       route = named_routes[helper_name]
-      routes.url_helpers.public_send(helper_method, *tokenized_params_for(route, params))
+      Routes.public_send(helper_method, *tokenized_params_for(route, params))
     end
 
     private
